Doing code cleanup, adding linter to project. Adding withdraw callback
