+++ ruff.toml
@@ -1,0 +1,1 @@
+line-length = 122
