--- src/ui/banner.py
@@ -1,6 +1,0 @@
-from kivy.uix.widget import Widget
-
-class Banner(Widget):
-    pass
-    # def __init__(self, **kwargs):
-    #     super(Subscription, self).__init__(**kwargs)
