+++ src/ui/currency.py
@@ -1,0 +1,8 @@
+from kivy.uix.dropdown import DropDown
+from kivy.uix.button import Button
+
+class CurrencyDropdown(DropDown):
+    pass
+
+class CurrencyButton(Button):
+    pass
